Moved filter function to utilities
